requirements.txt
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
lightgbm==4.3.0
shap==0.44.1
causalml==0.15.0
mlflow==2.11.1
fastapi==0.110.0
uvicorn==0.27.1
streamlit==1.32.0
evidently==0.4.22
psycopg2-binary==2.9.9
redis==5.0.1
faker==24.0.0
pytest==8.0.0
python-dotenv==1.0.1
lifelines==0.28.0
scikit-survival==0.22.2
dvc
python-dotenv>=1.0.0